+++ HR.LeaveManagement.Application/Features/LeaveRequest/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveRequest.Requests.Commands;
 using HR.LeaveManagement.Application.Persistence.Contracts;
+using HR.LeaveManagement.Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -23,6 +25,10 @@ namespace HR.LeaveManagement.Application.Features.LeaveRequest.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
         {
             var leaveRequest = await _leaveRequestRepository.Get(request.Id);
+            if (leaveRequest == null)
+            {
+                throw new NotFoundException(nameof(LeaveRequestt), request.Id);
+            }
             await _leaveRequestRepository.Delete(leaveRequest);
             return Unit.Value;
         }
